techdebt: move aurora-pg driver to the aurora-pg driver directory
for some reason the aurora postgres driver was living with the postgres
driver - instead of in the directory explicitly for the aurora postgres
driver.

this moves the aurora-pg driver into that directory so there's a clearer
delineation between the two drivers
